Add files via upload
The objective of the project is to build predictive model that will describe whether the patient is suffering from lever disease or not. The dataset contains the details of the patient like age, gender, blood tests which provide the details of Bilirubin, Protein, Sgpt, etc. In the project various hypothesis tests were accomplished to find the answer of question like 
does age or gender effects the liver of patients. What are the significant attributes to decide whether is patient possess
liver related ailments or not. 



Detials of the dataset


Age Age of the patient

Gender Gender of the patient

TB Total Bilirubin

DB Direct Bilirubin

Alkphos Alkaline Phosphotase

Sgpt Alamine Aminotransferase

Sgot Aspartate Aminotransferase

TP Total Protiens

ALB Albumin

A/G Ratio Albumin and Globulin Ratio

Selector field used to split the data into two sets (labeled by the experts) 1 Liver Patient, 2 non Liver Patient


The model so built provides accuray of 99% to predict whether patient is suffering from liver ailments

It was found that age and gender does not impacts the liver ailments of person using hypothesis test

Non parametric algorithm like Decision tree performs better to build the model as the data was not following normal distribution and outliers were there in the data

The model is ready for the deployment in the production